dextool: size_t is variable size depending on 32/64 bit architecture
Using ulong to make sure it is always 64 bit.

+++ plugin/graphml/test/integration.d
@@ -15,13 +15,12 @@ enum globalTestdir = "graphml_tests";
 
 /** Make a hash out of the raw data.
  *
- * Copied from the implementation
- * import cpptooling.utility.hash : makeHash;
+ * Copied from the implementation: dextool.hash : makeHash
  */
-size_t makeHash(T)(T raw) @safe pure nothrow @nogc {
+ulong makeHash(T)(T raw) @safe pure nothrow @nogc {
     import std.digest.crc;
 
-    size_t value = 0;
+    ulong value = 0;
 
     if (raw is null)
         return value;
